+++ server/api.py
@@ -9,8 +9,7 @@ from promptengineers.history import router as history_router
 from promptengineers.retrieval import router as retrieval_router
 from promptengineers.storage import router as storage_router
 from promptengineers.utils import logger
-
-from server.models.response import ResponseStatus
+from promptengineers.models.response import ResponseStatus
 
 
 app = FastAPI(