+++ server/config/test.py
@@ -1,4 +1,0 @@
-"""Config for test environment."""
-import os
-
-TEST_USER_ID = os.getenv("TEST_USER_ID", '000000000000000000000000')