+++ server/exceptions/__init__.py
@@ -1,3 +1,0 @@
-"""Exceptions for the server."""
-class ValidationException(Exception):
-	pass