+++ server/utils/functions.py
@@ -1,4 +1,0 @@
-
-def get_word_length(word: str) -> str:
-    """Returns the length of a word."""
-    return str(len(word))