+++ Changelog.md
@@ -7,5 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## v0.0.1
 ### Added 
+  - feature/add-promptengineers-pypi-package (2023-11-23)
   - feature/2-add-gpt-4-turbo
   - development